+++ lein2/resources/lein2/not_refreshed.clj
@@ -1,0 +1,3 @@
+(ns lein2.not-refreshed)
+
+(throw (Exception. "Code was loaded and shouldn't have been"))