+++ test-refresh/src/com/jakemccrary/test_refresh.clj
@@ -180,14 +180,14 @@
                                  (reset! printing true))]
       (if with-repl?
         (clojure.main/repl
-          :prompt #(do (while @printing (Thread/sleep 100))
-                       (clojure.main/repl-prompt))
-          :read (fn [request-prompt request-exit]
-                  (if-let [line (read-line)]
-                    (if-not (empty? (clojure.string/trim line))
-                      (read-string line)
-                      (do (run-test-refresh!) :run-tests))
-                    (System/exit 0))))
+         :prompt #(do (while @printing (Thread/sleep 100))
+                      (clojure.main/repl-prompt))
+         :read (fn [request-prompt request-exit]
+                 (if-let [line (read-line)]
+                   (if-not (empty? (clojure.string/trim line))
+                     (read-string line)
+                     (do (run-test-refresh!) :run-tests))
+                   (System/exit 0))))
         (loop [c (.read System/in)]
           (if (= c -1)
             (System/exit 0)
@@ -224,10 +224,16 @@
         run-once? (:run-once options)
         with-repl? (:with-repl options)
         watch-dirs (:watch-dirs options)
+        refresh-dirs (:refresh-dirs options)
         monitoring? (atom false)]
+
+    (when (seq refresh-dirs)
+      (println "Only refreshing dirs:" (pr-str refresh-dirs))
+      (apply clojure.tools.namespace.repl/set-refresh-dirs refresh-dirs))
 
     (when report
       (println "Using reporter:" report))
+
     (when (:quiet options)
       (defmethod capture-report :begin-test-ns [m]))
 
@@ -239,7 +245,9 @@
             (reset! keystroke-pressed nil)
             (reset! printing true)
 
-            (when (and with-repl? @monitoring? something-changed?) (println))
+            (when (and with-repl? @monitoring? something-changed?)
+              (println))
+
             (print-banner)
 
             (let [was-failed (tracking-failed-tests?)
@@ -272,6 +280,4 @@
           (recur (dissoc new-tracker
                          :clojure.tools.namespace.track/load
                          :clojure.tools.namespace.track/unload))
-          (do
-            (println "K THX BYE")
-            (System/exit 0)))))))
+          (System/exit 0))))))